// src/app/services/API/models/new-instrument-dto.ts
/* tslint:disable */
export interface NewInstrumentDto {
    boardId?: number;
    boardName?: string;
    id: number;
    isActive: boolean;
    isInBourse: boolean;
    marketId?: number;
    marketName?: string;
    name: string;
    nameEn?: string;
    symbol?: string;
    symbolEn?: string;
    ticker: string;
    type: number;
}